[Feat] : 월별통계 조회

+++ src/main/java/ETA/whats_your_eta/api/domain/statistics/VisitedRegion.java
@@ -25,5 +25,5 @@ public class VisitedRegion {
 
     private String region;
 
-    private Float percentage;
+    private Double percentage;
 }
